+++ sqleto_annotation/lib/sqleto_annotation.dart
@@ -4,3 +4,4 @@ export 'src/sqleto_annotations.dart';
 export 'src/sqleto_type.dart';
 export 'src/sqleto_default_value.dart';
 export 'src/sqleto_validator.dart';
+export 'src/sqleto_schema.dart';